fix(release): extract runtime platform dependencies [skip release]
